Fix e^x-1 for large input values.
Specifically values that are too large to represent in the packed form but
which can be represented in the real_t form.  The increase in exponent range
of real_t values triggered problems in the correction code for inputs around
zero.  The solution is to not both with the correction for inputs not near
zero.  This will speed up e^x-1 computations most of the time.

The choice of the interval (-.1, .1) is arbitrary.  This range loses at most
one significant digit due to cancellation.  The interval (-1, 1) loses a third
of a digit but is ten times wider.