Created two new number theory commands.
Building on FACTORS in X.FN, I have expanded X.FN with a new page containing
two new entries:  EVPFACTS and EULPHI.

These are similar to commands available in the HP 50G.  EVPFACTS is the inverse
of the FACTORS command, like the EVAL key on the 50G does.  It expands out the
factors from the matrix and produces a long integer with the given factorisation.

EULPHI computes the Euler Phi function of a long integer, outputting another
long integer.  It uses FACTORS to compute the prime factorisation and use that
in the calculation.