Restore numbers more reliably
When restoring non-user numbers from the save state file, force base 10
for input.  This fixes the problem with registers which are numbered with
three digits which can include leading zeros (which used to force octal
interpretation).

Also change the text to number conversion functions to use strtol(3),
strtoul(3), strtoll(3) and strtoull(3).